Testing

+++ server/controllers/aboutUserController.js
@@ -1,24 +1,28 @@
 import AboutUser from "../model/aboutModel.js";
-import jwt  from "jsonwebtoken";
+import jwt from "jsonwebtoken";
 
-const aboutUserPost = async (req , res) => {
-    try {
-         const { aboutUser } = req.body
+const aboutUserPost = async (req, res) => {
+  try {
+    if (req.cookies.jwt) {
+      const { aboutUser } = req.body;
 
-         const token = await req.cookies.jwt;
-         const decoded = jwt.verify(token , process.env.SECRET_KEY)
-         const userId = await decoded.id;
+      const token = await req.cookies.jwt;
+      const decoded = jwt.verify(token, process.env.SECRET_KEY);
+      const userId = await decoded.id;
 
-         await AboutUser.deleteOne({userId: userId})
+      await AboutUser.deleteOne({ userId: userId });
 
-         const about = await AboutUser.create({about: aboutUser , userId: userId})
+      const about = await AboutUser.create({
+        about: aboutUser,
+        userId: userId,
+      });
 
-         res.json({message: "Updated successfully"})
-        
-    } catch (err) {
-        res.status(500).json({message: "internal error"})
-        console.log(err)
+      res.json({ message: "Updated successfully" });
     }
-}
+  } catch (err) {
+    res.status(500).json({ message: "internal error" });
+    console.log(err);
+  }
+};
 
-export default aboutUserPost
+export default aboutUserPost;
